Status.hs: fix handling of ambiguou hackage packages
I've noticed it on output of 'hackport status --from-hackage':
today's Cabal-1.24.1.0 was missing in the output

The root cause here is use of 'Data.Map.fromList' on a list
with duplicate 'dev-haskell/cabal'  key resulted from
2 distinct hackage packages: 'Cabal-1.24.1.0' and 'cabal-0.0.0.0'
where the latter was preferred.

The fix is simple: use 'Data.Map.fromListWith' passing
explicit merge strategy.

Signed-off-by: Sergei Trofimovich <siarheit@google.com>