Metadata: automatically generate flags for metadata.xml
This patch changes the type signature and behaviour of makeDefaultMetadata
and mergeEbuild to accomodate the automatic generation of Cabal flags and
their descriptions in the metadata.xml.

The changes have been tested on two packages: one which has USE flags
(app-text/pandoc), and one which does not (dev-haskell/control-monad-free).
In both circumstances, the generated metadata.xml is correctly formatted.

Where the existing and new metadata.xml differ, the current behaviour of
HackPort remains the same, ie, the current metadata.xml will not be
overwritten.

In the case of a package with no USE flags, currently this patch still writes
the <use> and </use> tags, but leaves its contents empty.

Bug: https://github.com/gentoo-haskell/hackport/issues/35
Signed-off-by: Jack Todaro <jackmtodaro@gmail.com>