Merge, Portage.PackageId: Add diff functionality
This teaches hackport to generate and display a diff between a newly-merged
version of a package and its most immediate previous version, if available.
This is not currently switchable. If there is no previous version available
such as when merging a new package, hackport will not call diff.

The main benefit of this functionality is to allow the ebuild author to more
easily detect unintended changes between version bumps. A small diff is
likely to suggest no major changes, whereas a large diff should signal to
the ebuild author that further investigation may be necessary.

For example:

    foo-a.b.ebuild has been patched using cabal_chdep. Hackport has no way
    of knowing if version foo-a.b+1.ebuild also requires this. The ebuild
    is likely to pass repoman, but it is unlikely to emerge as intended.

    Package bar has its dependency on cabal removed to avoid circular
    dependencies (this is particularly relevant to packages close to GHC).
    Hackport has no way of knowing if this is still necessary. The new version
    of the ebuild may pass repoman, but it may cause circular dependencies
    for users.

Signed-off-by: Jack Todaro <jackmtodaro@gmail.com>