Portage/Host.hs: improved portage guess
Found a few kinks while getting started with hackport (as of 0.6 release).

Problem: The approach to grab `portage_dir` and `overlay_list` results in
`Nothing`, as they are no longer reported by `emerge --info`. There are error
messages that advise to set PORTDIR_OVERLAY in '/etc/make.conf'. Wrong file
location and obsolete variable. The message in (aptly named)
`showAnnoyingWarning` reflects missing information (my portage is in
'/var/portage' but it prints '/usr/portage'). This applies to `getInfo` using
emerge (no idea for paludis but the adjustments should not affect it).

Solution: Query `portageq` which is bundled with sys-apps/portage (instead of
parse-and-pray). Note that it does not have an advisable direct equivalent of
PORTDIR_OVERLAY, but provides the path to specific repositories. Assumption
made here: the repository of interest goes by the name "haskell". Updated two
related error messages in "Error.hs". Eventually one might remove altogether
the overlay_list and replace it with a simple and direct repo_dir.

Tested with a patched ebuild of hackport-0.6.
ghc 8.6.5, portage 2.3.76

Signed-off-by: Philippe Baril Lecavalier <pbl.ltx@gmail.com>