Portage/PackageId.hs: use parseCabalPackageName instead of parsec
Just as in a previous commit, we need to use parseCabalPackageName
to parse the Cabal PackageName type, since the 'parsec' method
immediately fails on input of a version number.

Given that we frequently parse package names with their version appended,
we need to use our in-house parser which backtracks upon reaching
a version number. In the case of the bug referenced below, this caused
issues with parsing packages available in the Portage tree (::gentoo)
and the haskell overlay, which is used in the 'status' and 'list' commands.

Bug: https://github.com/gentoo-haskell/hackport/issues/63
Reported-By: Philippe Baril Lecavalier

Signed-off-by: Jack Todaro <jack.todaro@posteo.net>