Portage/Metadata.hs, Merge.hs: automate metadata.xml USE flag handling
The purpose of this change is to automate as much as possible
the generation of a correct metadata.xml, so that less manual
intervention is required by new and existing contributors.

Prior to this commit, if a package adds new configuration
flags in a new version, these flags are not added as USE
flags to the metadata.xml. Hopefully, if a contributor uses
repoman, they will be alerted to this issue and know how to
fix it. In the worst case, they simply 'git commit' with no repoman
checks, resulting in a PR failing checks and a request to
make changes (or the maintainer fixes it themselves), or the
contributor ends up overwhelmed at not understanding the repoman
failure and steps away from becoming a potential contributor.

This commit changes hackport's behaviour in the following way:

  1. If there is no existing metadata.xml, behaviour is unchanged
  2. If there is an existing metadata.xml identical to the newly
     generated metadata.xml, behaviour is unchanged
  3. If there is an existing metadata.xml which differs from the
     newly generated metadata.xml:
       a) in its description, then the new description is used
       b) in its USE flag descriptions (but no new flags added),
          then the new USE flag descriptions are used. HackPort
          will alert the user to this change at the standard
	  verbosity
       c) in that it contains new USE flags, then these flags
          are added to the USE flags in the metadata.xml

In short, if the new and existing metadata.xml differ, the new
one always takes precedence. HackPort will report some of these
differences to the user at the standard level of verbosity.

All in all, this will generate a valid metadata.xml more often
than HackPort did prior to this commit. As such this behaviour is
not currently switched. HackPort will be smart enough to only add
active USE flags, respecting the 'hackport: flags:' variable and
preserving any and all name mangling, so there should be little reason
to switch this behaviour off.

For a deeper explanation of some of the changes:
- We now store USE flags as a 'Data.Map.Strict.Map String String',
  and all ancillary functions have been updated to handle this change
- The 'parseMetadata' function can now extract USE flags and their
  descriptions from a provided 'Text.XML.Light.Element'.
- We have a 'pureMetadataFromFile' function which now drives the impure
  'metadataFromFile' function, which converts a supplied 'Data.Text.Text'
  into a 'Maybe Metadata'.
- We now use Data.Text strings instead of ByteStrings for IO. This is to
  preserve UTF-8 encoding in the correct manner (previous usage of ByteString
  builders is supposedly not the best way to handle UTF-8).

Signed-off-by: Jack Todaro <solpeth@posteo.org>