tests: add hspec tests, more doctests
The main change in this commit is the addition of a hspec
test-suite. As stated in the initial doctests commit, doctest
was added primarily to check documentation examples and properties,
despite the added benefit of free regression testing of some sort.

This commit adds proper unit/property testing using hspec. The
benefit over simply using doctest is being able to write more
complex test cases without cluttering the code, as well as
allowing us to operate in modules which rely on Paths_hackport,
such as Portage/EBuild.hs.

A test has been written to check for hackport issue #36, which
passes the test and therefore appears to be fixed (although I do not
know the reason). Should the issue pop up again, the test-suite
should catch it.

Additionally, two of the three existing HUnit tests have been converted into hspec
tests, to improve integration and reduce the number of active test-suites
down to two: doctest and hspec. The only existing HUnit test which remains
unported is resolveCat.hs, which fails in a sandboxed environment anyway
and has therefore been disabled. Overall, these changes should allow us
to re-enable tests in the hackport ebuild in ::haskell.

The HUnit tests have not yet been removed from the .cabal file or the
source repository, to allow time for anyone to check that the porting is
correct (which it should be, since the porting was largely copy-pasting
from the HUnit tests into a Spec file).

There are currently 28 tests in the test-suite, and hopefully more
will be added over time.

Bug: https://github.com/gentoo-haskell/hackport/issues/36
Signed-off-by: Jack Todaro <solpeth@posteo.org>