Portage/Cabal.hs: fix licence version translation
I was unaware that 'GPL-2.0' was an invalid licence in Gentoo,
but this is what Cabal returns since switching to SPDX licence
identifiers. For example, try 'hackport merge pandoc'.

This is a quick fix tackling the GPL case, and the test suites
have been updated to catch this. I'll add more tests for this
in future commits.

Signed-off-by: Jack Todaro <solpeth@posteo.org>