Portage/Cabal.hs: fix GNU licence handling, handle 'or-later'
We previously didn't handle 'or-later'-style GNU licences,
such as GPL-3+. Since Cabal now internally uses SPDX licences,
we can now differentiate between GPL-3 and GPL-3+, as well as
the AGPL-3 and 3+ and the LGPL-3 and 3+, but not LGPL-2.0 and
LGPL-2.1 and their 'or-later' versions, since Cabal does not
internally differentiate them. (We would need to port our
functions to directly manipulate SPDX licences rather than
Cabal licences).

This also updates the test suite to check for the correct behaviour.
Cabal seems inconsistent in the way it transforms GNU SPDX
licences into Cabal licences, so if the behaviour changes in
the future the regression should be caught.

Signed-off-by: Jack Todaro <solpeth@posteo.org>