Portage/Metadata.hs: make makeDefaultMetadata return a Text
This avoids packing the String elsewhere. Since we shouldn't
need to deal with a metadata String instead of a metadata
Text, we may as well pack the String directly in the function.

This commit also adds more tests to ensure correct metadata.xml
generation and USE flag entry formatting. Some of these tests
are quite precise, so any change in formatting (valid or not)
will fail some of these tests. On the other hand we can guarantee
that our functions will continue to produce the same metadata.xml.

Signed-off-by: Jack Todaro <solpeth@posteo.org>