Portage/Version.hs: fix Suffix parser, add more property tests
New test cases detected a parser error for the Suffix parser. This
commit has fixed this.

Outstanding is a bug causing the Version parser to fail
if an optional version char is parsed. Currently this bug appears
to be due to Portage.PackageId.parseCabalPackageName trying to
parse a version number with no dot (.) separators.

Signed-off-by: Jack Todaro <solpeth@posteo.org>