Portage/PackageId.hs: rename variables to avoid name shadowing
category -> cat, in three places.

The 'parseCabalPackageName' parser has also been extended to ensure
that it can parse a single '+' character as a valid part of a
package name, as defined by the Gentoo ebuild naming spec.

Removed was a comment detailing a suspected bug in this parser,
due to that suspicion being incorrect.

Signed-off-by: Jack Todaro <solpeth@posteo.org>