tests: fix Arbitrary ComplexVersion instance
The previous ComplexVersion instance of Arbitrary was capable
of generating invalid Versions. Specifically it defied the Gentoo
ebuild naming specification (see
https://devmanual.gentoo.org/ebuild-writing/file-format/index.html)
in that it allowed a single version number and a single version char
to coexist, e.g:

'foo-bar1-1a_pre-r1'.
          ^^

This version is invalid, since '1a' is ambiguous. It is a valid package
name, which according to the Gentoo spec takes precedence over such use
as a version number. And previously, ComplexVersion erroneously treated
this as a valid Version, too. This means that optional version chars are only
allowed if the version remains unambiguous, i.e. that the version number
contains dot (.) separators to clearly distinguish itself as a version.

Therefore the previous test failures were not a result of a faulty parser,
rather a result of a faulty instance of Arbitrary. This instance has been
fixed, with the added benefit of now being able to produce Versions with
the version char set to Nothing, as well as version numbers with only one
digit (previously constrained to producing at least two digits to work
around this buggy instance).

Property tests have also been adjusted to test package names with plus
(+) signs, to test the operation of the updated 'parseCabalPackageName'
parser which can now parse plus signs, as per the Gentoo spec.

Signed-off-by: Jack Todaro <solpeth@posteo.org>