Portage/Metadata.hs: generate a Gentoo CI-valid metadata.xml
The previous commit addressed Gentoo's CI bot complaining about
an empty <use> element. This commit addresses the
RedundantLongDescription warning, which is triggered if
the long description is either equal to the ebuild's DESCRIPTION
field, or it is less than or equal to 150 characters in length.

This commit therefore adjusts hackport's behaviour to check for
a valid long description before writing it, and it does so in
two places. Firstly, in Portage/Metadata.hs in the
makeDefaultMetadata function, we now check for an empty long
description field or a long description below the required length.
Secondly, in Merge.hs, when invoking makeDefaultMetadata we pass
an empty long description string if it is equal to the ebuild's
DESCRIPTION field, otherwise we pass the long description as-is.

This behaviour can be reworked to use the Maybe monad to better
describe why we are passing an empty string, but for now this
works. An attempt at using Maybe, while it worked, also introduced
more complexity than should be necessary.

Therefore, the checks are carried out in two places. If when
invoking makeDefaultMetadata the long description is equal to
the ebuild's DESCRIPTION, then we pass an empty string to
makeDefaultMetadata. If makeDefaultMetadata is fed an empty
long description string, it does not write a <longdescription>
element. It will also neglect writing a <longdescription> element
if the long description passed to it does not exceed 150 characters
in length.

This should hopefully help to avoid introducing new QA issues into
::gentoo when it comes to the format of the metadata.xml, which was
a significant issue during the recent syncing between ::haskell and
::gentoo of around 300 packages - many of which had metadata.xml files
that introduced new warnings into ::gentoo's CI.

The spec test suite has also been updated to check for this new behaviour,
bringing the number of hspec tests up to 57.

Signed-off-by: Jack Todaro <solpeth@posteo.org>