tree-wide: warn user if bumping a ::gentoo package
This commit teaches hackport to advise the user to also bump the
package for ::gentoo, if the package being bumped is on the list
kept in Merge.Utils.gentooPackages and if we are not operating
hackport inside ::gentoo. Note that in no way is this list definitive
and I invite addition to it, but the packages currently on the list
are those which were targeted in the recent ::haskell to ::gentoo
package sync and therefore serve as a useful starting point.

The intention is to help with keeping ::gentoo and ::haskell in sync
for a handful of commonly-used packages, which means this commit
only makes sense if the Gentoo Haskell project decides to treat
::gentoo haskell maintenance in this way. The separation between
the two repositories would thus become ::gentoo for common haskell
packages and their dependencies, and ::haskell as a testing ground for
the bleeding edge of the haskell package ecosystem.

Note that this could have been implemented to check if the package being
bumped exists in ::gentoo at all rather than maintaining a manual list
of packages, but I have avoided this for two reasons:

  1) More frequents bumps can mean more frequent breakages, which is
     not what we want in ::gentoo, but is something more tolerable
     (albeit still undesirable) in ::haskell
  2) Not every package that is currently in ::gentoo may belong in ::gentoo.
     Maintaining a hand-picked list of packages to maintain in ::gentoo
     may help to refine and define the scope of each repository, with that
     list being subject to change, demand of the user base and opinion
     of the maintainers.

This simple commit therefore makes the assumption that the package being
warned about does not yet exist in ::gentoo. This assumption is made on the
basis that ::haskell is the testing ground from which ::gentoo draws its
haskell packages, therefore always ::haskell -> ::gentoo, not the other way
round.

GHC and its core libraries are in gentooPackages for the convenience of
seeing which packages require bumping with ghc, but they remain commented out
since those packages require particular care and coordination when bumping
and are not suited to ad-hoc version bumping unlike the popular packages
pandoc, hlint and stack.

Signed-off-by: Jack Todaro <solpeth@posteo.org>