Merge.hs: parallelise expensive list computation
This implements a partial mitigation of the slowdown experienced
in the bug referenced below, but it does not address the space
leak.

On a 2-core, hyperthreaded machine running on four 'cores', computing
all possible flag combinations takes about a third of the time that the
single-threaded version takes. This has been achieved with a simple
one-liner in Merge.hs and a number of NFData instances to enable
deep evaluation of the data involved.

Computing the dependency strings for the ebuild is proving difficult
to parallelise and remains a significant, single-threaded bottleneck.
In the case of xmobar, it takes roughly four times as long as the
flag combinations computation to finalise the dependency strings for
writing. This is something that requires further investigation.

In any case, this patch warns the user that hackport may take a while
to complete if there are at least 12 cabal flags in the package. For
a package like git-annex with around 90 dependencies and 8192 possible
flag combinations (that's 13 package flags), this means waiting around
two minutes for the computation to complete on my machine. For xmobar,
with half the number of dependencies the same computation is about 32
seconds.

There are very few packages that will run into this problem, thankfully.
But packages such as git-annex and xmobar are frequently updated, so
speeding this process up with parallelism will free up time for other
maintenance tasks and use hardware more efficiently.

Bug: https://github.com/gentoo-haskell/hackport/issues/56

Signed-off-by: Jack Todaro <solpeth@posteo.org>