Merge.hs: avoid re-using an existing DESCRIPTION for new ebuilds
Carrying over an old ebuild's DESCRIPTION to a new ebuild is lossy:
It does not correctly preserve the DESCRIPTION string if it has
escaped characters in it.

There may be a fix available in Portage/EMeta.hs, but this commit instead
avoids using the functionality entirely on the basis that if a
package updates its DESCRIPTION upstream, we should probably honour
that. This is similar to how we update the <longdescription> element
in the metadata.xml if the upstream package maintainer updates it,
which is important for keeping it current, valid and relevant to
the package's most recent version.

Another test case specifically testing contravariant-extra's
quoted DESCRIPTION string has been added. Hopefully test cases
for Portage/EMeta.hs will be forthcoming in the future.

Closes: https://github.com/gentoo-haskell/hackport/issues/36
Signed-off-by: Jack Todaro <solpeth@posteo.org>