Portage/Ebuild.hs: generalise toHttps function
pkgcheck has the useful feature of validating all URIs listed in
an ebuild and checking if they have a https counterpart. This has
uncovered that the ebuildTemplate function was writing a default
homepage as http://hackage.haskell.org, without encryption. This
has now been updated since hackage.haskell.org has https support.

The toHttps function has also been somewhat generalised: it now
parses a supplied String as a URI, determines if the uri scheme
is http: and checks if the uri is an element of httpsHomepages. If
so it replaces http: with https:, if not it returns the supplied
String as-is.

We can add to this list whenever we come across homepages that are
https-aware, but not listed as such by the package author.

Signed-off-by: Jack Todaro <solpeth@posteo.org>