Portage/EBuild.hs: switch from https whitelist to http whitelist
The previous implementation of 'toHttps' used a https whitelist,
meaning that only urls contained in the whitelist would be converted
to https. This has a couple of issues:

  1. Changes are made to hackport only to add https awareness of a
     certain website
  2. As more packages are added to hackage, we will need to grow the
     https whitelist
  3. If hackport maintenance doesn't keep up, more ebuilds will be
     generated with http-only homepages where the homepages are in fact
     https-aware. The longer hackport maintenance were to stall, the more
     ebuilds would be generated with http-only homepages, causing CI
     issues with pkgcheck and Gentoo CI.

But in modern times (largely because of Let's Encrypt) many websites
are https-aware despite this not being reflected in the .cabal file
of a package. It should be safe to assume that a website is https-aware
until proven otherwise.

Thus, the https whitelist has been converted into a http whitelist. The
benefits are:

  1. This list should *shrink* over time (after we initially determine the key
     packages which genuinely do not have https-aware homepages) as more websites
     become https-aware
  2. Assuming the scenario of hackport maintenance waning or entirely stalling,
     correctness should *increase* with time as regards the generation of
     homepages (based on the current trends of https awareness in websites, and
     assuming the scenario where we have not yet determined all http-only homepages).
  3. Fewer changes to the whitelist will have to be made compared to the current
     https whitelist, reducing maintenance burden and freeing up time for other
     tasks.

Additionally 'toHttps' has been split out as a top-level function,
and some test cases have been added to the spec test-suite to
ensure its correctness.

Signed-off-by: Jack Todaro <solpeth@posteo.org>