Portage/Metadata.hs: don't add global USE flags to metadata.xml
Strip global USE flags from the flags Map, prior to writing them
into the metadata.xml. Note that this is currently implemented in
a way which avoids parsing the use.desc file, and instead relies
upon manually-designated USE flags to treat as global in all haskell
packages.

Tests added to the spec test-suite.

Signed-off-by: Jack Todaro <solpeth@posteo.org>