Merge/Utils.hs: rewrite drop_prefix function
Rewritten to be less dependent on hard-coded prefix
lengths, and to be more dynamic.

It now directly recognises and drops any prescribed
prefixes in the prefixes list, making it more extensible
and robust should further prefixes be added. Rather than having
to add a prefix, then add a case to recognise said prefix, and then
drop the prefix based on a hardcoded number of elements to drop,
we can just add the prefix string to the prefixes list, and the function
will do the rest.

The test suite has also been updated to reflect this change, as well
as to make some test cases more useful and relevant with more unit
testing.

Signed-off-by: Jack Todaro <solpeth@posteo.org>