Portage/Metadata.hs: correctly handle special XML characters in flags
Previously, special XML characters such as '<' were not correctly
transformed to avoid XML errors. Using the 'xml' package, this is
now properly handled in flag names and descriptions to avoid these
issues cropping up. See for instance 'hackport merge hsopenssl-0.11.7'
before and after this patch.

Test cases have been added to the spec test-suite to check for correct
XML handling, as well as for correct formatting when multiple flags
are added to the metadata.xml.

Signed-off-by: Jack Todaro <solpeth@posteo.org>