Cabal2EBuild,Ebuild.hs: pull in hackage revisions
These changes read the x-revision value of the
package's PackageDescription pulled from Hackage
when performing a hackport merge.

If the key exists, the value is applied to the
CABAL_HACKAGE_REVISION line of the ebuild. If
it does not exist, the default value of 0 is
used.

It was decided at this stage to always write
the CABAL_HACKAGE_REVISION line for easy updating
later on, however I am unopposed to having it written
conditionally based upon the existence of revisions.

Signed-off-by: Jack Todaro <solpeth@posteo.org>