Add remote-id support
Remote ids (from hackage, github and other sources) are now
automatically added to metadata.xml when using the merge and
make-ebuild commands.

There is now a --not-on-hackage flag for the make-ebuild command. This
prevents the "hackage" remote-id from being written to metadata.xml.

The new Env monad has been implemented in various places.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>