Switch to optparse-applicative
This is a major rewrite of Main.hs, switching to the
optparse-applicative package and placing much of the logic in
src/Hackport/. The command-line semantics are largely the same, although
there may be some subtle differences because of the switch to a new
parser library.

It also introduces a new monad which will carry global and
command-specific environment state:

type Env env = ReaderT (GlobalEnv, env) IO

This should help facilitate passing state from the command-line deep
into the program.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>