Handle -Wincomplete-uni-patterns
This warning is turned on by -Wall in ghc-9.2 and will cause build
failures when the 'pedantic' flag is used. It is being explicitly
enabled in hackport.cabal so the behavior is consistent across ghc
versions.

For now, it is simply being disabled in the few modules where it pops up,
but things should be rewritten in the future so we don't have to disable
this warning.

Closes: https://github.com/gentoo-haskell/hackport/issues/102
Signed-off-by: hololeap <hololeap@users.noreply.github.com>