Add shell completion functionality
Use optparse-applicative's shell completion functionality

The completion for the 'merge' subcommand involves searching the full
list of package-version strings available on Hackage for a prefix. To
help facilitate this, a Trie datatype has been added.

The generated Trie is cached in `~/.hackport/packages.cache`. The Trie
file is removed when `hackport update` is called, forcing the cache to
be rewritten on next use.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>