Revert "Portage/GHCCore.hs: Make process an upgradeable package"
Exposing the 'process' package as a portage package has proven to be
a bad idea, creating many complications and system inconsistencies.

The version that comes bundled with GHC needs to be the only version
on the system. If upgrades to the 'process' package are needed, it
needs to be done through dev-lang/ghc.

This reverts commit f5b0ae67df18014585605825765ce76dc7ac00fb.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>