Only add current ARCH to KEYWORDS
The assumption that amd64 and x86 always work is erroneous. For one,
there have been observable failures on x86 which do not appear in amd64.
Also, it is a safer assumption that the packager is testing on their
current ARCH rather than on amd64 and x86.

Passing previous KEYWORDS forward is not a sound idea. It is possible
that the current maintainer is on a different ARCH than the previous
one. The most sane way to handle KEYWORDS is to only add the current
ARCH.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>