Portage.EBuild: Switch to LinesBuilder
LinesBuilder is a custom Monad that works by building up a list of
prettyprinter Docs and then concats them with vcat. It is helpful for
when a logical piece of the ebuild may use a variable number of lines,
including no lines at all (for instance when no CABAL_HACKAGE_REVISION
or CABAL_PN are used).

Improve some formatting

Signed-off-by: hololeap <hololeap@protonmail.com>