Display summary of warnings at end of execution
Hackport is noisy. At the end of the output, display any warnings that
were fired so that they are easier to notice. We use a DList to hold
the list of warnings. This adds `StateT WarningBuffer` to the Env monad
stack and runs displayWarnings (new function) as the last part of
runEnv.

Add `warn` function to the Util module and display warnings whenever
`die` is called.

Convert `errorWarnOnUnbuildable` in the Merge module to use the new
warning functionality.

Signed-off-by: hololeap <hololeap@protonmail.com>