Merge/Dependencies: Ignore hsc2hs completely
This is a utility that ships with GHC, but there is no haskell package
for it (either on hackage or on the system.) It is occasionally
referenced under 'build-tool-depends' in the cabal file.
(Example: lukko-0.1.2)

References buildToolsProvided which is a list that includes hsc2hs,
defined in the same module.

Closes: https://github.com/gentoo-haskell/hackport/issues/134
Signed-off-by: hololeap <hololeap@protonmail.com>