Add support for "#hackport: external-depends: ..."
This is intended for situations where a Haskell package depends on a
non-Haskell executable (which does not have a canonical field in the
.cabal file spec).

Does not currently support flags or version ranges on the atoms. Does
not support BDEPEND yet as this needs to be added to
Merge.Dependencies.EDep.

Random example:

    #hackport: external-depends: rdepend:dev-db/postgresql,depend:app-misc/ca-certificates

Signed-off-by: hololeap <hololeap@protonmail.com>